requirements.txt
dill==0.3.6
Django==4.1.6
gensim==4.2.0
jieba==0.42.1
matplotlib==3.5.3
numpy==1.22.3
openpyxl==3.1.0
Pillow==9.4.0
pymysql==1.0.2
scikit_learn==1.2.1
tensorflow==2.11.0
torch==1.12.0
torchtext==0.13.0
wordcloud==1.8.2.2
